Move ProjectDependencyFactory to projects module
