+++ collect_app/src/main/java/org/odk/collect/android/instancemanagement/InstancesDataService.kt
@@ -13,7 +13,6 @@ import org.odk.collect.android.instancemanagement.autosend.InstanceAutoSendFetch
 import org.odk.collect.android.instancemanagement.autosend.getAutoSendMode
 import org.odk.collect.android.notifications.Notifier
 import org.odk.collect.android.openrosa.OpenRosaHttpInterface
-import org.odk.collect.android.projects.ProjectDependencyFactory
 import org.odk.collect.android.projects.ProjectDependencyModule
 import org.odk.collect.android.utilities.ExternalizableFormDefCache
 import org.odk.collect.android.utilities.FormsUploadResultInterpreter
@@ -21,6 +20,7 @@ import org.odk.collect.androidshared.data.AppState
 import org.odk.collect.forms.Form
 import org.odk.collect.forms.instances.Instance
 import org.odk.collect.metadata.PropertyManager
+import org.odk.collect.projects.ProjectDependencyFactory
 import java.io.File
 
 class InstancesDataService(
