+++ collect_app/src/main/java/org/odk/collect/android/utilities/ChangeLockProvider.kt
@@ -1,6 +1,6 @@
 package org.odk.collect.android.utilities
 
-import org.odk.collect.android.projects.ProjectDependencyFactory
+import org.odk.collect.projects.ProjectDependencyFactory
 import org.odk.collect.shared.locks.ChangeLock
 import org.odk.collect.shared.locks.ReentrantLockChangeLock
 import javax.inject.Singleton
